Fix a bug in the SBC object which launched the firmware window instead of the shutdown/reboot window when pressing the shutdown/reboot button.
Fix for a warning in the MJDVacuumController.


git-svn-id: svn://orca.physics.unc.edu/Orca64@12088 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/MJDVacuum/ORMJDVacuumController.m
@@ -468,7 +468,7 @@
 
 - (IBAction) constrainingGaugeAction:(id)sender
 {
-    int index = [sender indexOfSelectedItem];
+    long index = [sender indexOfSelectedItem];
     if(index == 0 || index == 2) [model setConstrainCryostatGauge:YES];
     else [model setConstrainCryostatGauge:NO];
     if(index == 1 || index == 2) [model setConstrainRGAGauge:YES];
